Fixed TODOs

+++ core/impl/src/main/java/com/blazebit/persistence/impl/PaginatedCriteriaBuilderImpl.java
@@ -69,8 +69,8 @@ public class PaginatedCriteriaBuilderImpl<T> extends AbstractQueryBuilder<T, Pag
         Query idQuery = em.createQuery(idQueryString);
         parameterizeQuery(idQuery);
 
-        List ids = idQuery.setFirstResult((int) firstRow)
-            .setMaxResults((int) pageSize)
+        List ids = idQuery.setFirstResult(firstRow)
+            .setMaxResults(pageSize)
             .getResultList();
         
         if (ids.isEmpty()) {
